+++ crates/stackable-operator/src/builder/configmap.rs
@@ -1,6 +1,14 @@
-use crate::error::{Error, OperatorResult};
 use k8s_openapi::{api::core::v1::ConfigMap, apimachinery::pkg::apis::meta::v1::ObjectMeta};
+use snafu::{OptionExt, Snafu};
 use std::collections::BTreeMap;
+
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+#[derive(Snafu, Debug)]
+pub enum Error {
+    #[snafu(display("object is missing key {key:?}"))]
+    MissingObjectKey { key: &'static str },
+}
 
 /// A builder to build [`ConfigMap`] objects.
 #[derive(Clone, Default)]
@@ -41,12 +49,13 @@ impl ConfigMapBuilder {
         self
     }
 
-    pub fn build(&self) -> OperatorResult<ConfigMap> {
+    pub fn build(&self) -> Result<ConfigMap> {
+        let metadata = self
+            .metadata
+            .clone()
+            .context(MissingObjectKeySnafu { key: "metadata" })?;
         Ok(ConfigMap {
-            metadata: match self.metadata {
-                None => return Err(Error::MissingObjectKey { key: "metadata" }),
-                Some(ref metadata) => metadata.clone(),
-            },
+            metadata,
             data: self.data.clone(),
             ..ConfigMap::default()
         })