+++ crates/stackable-operator/src/builder/pod/volume.rs
@@ -12,7 +12,7 @@ use snafu::{ResultExt, Snafu};
 use tracing::warn;
 
 use crate::{
-    builder::ObjectMetaBuilder,
+    builder::meta::ObjectMetaBuilder,
     kvp::{Annotation, AnnotationError, Annotations},
 };
 
@@ -262,7 +262,7 @@ impl VolumeMountBuilder {
     }
 }
 
-#[derive(Debug, Snafu)]
+#[derive(Debug, PartialEq, Snafu)]
 pub enum SecretOperatorVolumeSourceBuilderError {
     #[snafu(display("failed to parse secret operator volume annotation"))]
     ParseAnnotation { source: AnnotationError },
@@ -414,7 +414,7 @@ impl ListenerReference {
 
 // NOTE (Techassi): We might want to think about these names and how long they
 // are getting.
-#[derive(Debug, Snafu)]
+#[derive(Debug, PartialEq, Snafu)]
 pub enum ListenerOperatorVolumeSourceBuilderError {
     #[snafu(display("failed to convert listener reference into Kubernetes annotation"))]
     ListenerReferenceAnnotation { source: AnnotationError },
@@ -426,9 +426,9 @@ pub enum ListenerOperatorVolumeSourceBuilderError {
 ///
 /// ```
 /// # use k8s_openapi::api::core::v1::Volume;
-/// # use stackable_operator::builder::ListenerReference;
-/// # use stackable_operator::builder::ListenerOperatorVolumeSourceBuilder;
-/// # use stackable_operator::builder::PodBuilder;
+/// # use stackable_operator::builder::pod::volume::ListenerReference;
+/// # use stackable_operator::builder::pod::volume::ListenerOperatorVolumeSourceBuilder;
+/// # use stackable_operator::builder::pod::PodBuilder;
 /// let mut pod_builder = PodBuilder::new();
 ///
 /// let volume_source =