+++ crates/stackable-webhook/src/tls.rs
@@ -46,6 +46,31 @@ pub enum Error {
     },
 }
 
+/// Custom implementation of [`std::cmp::PartialEq`] because some inner types
+/// don't implement it.
+///
+/// Note that this implementation is restritced to testing because there are
+/// variants that use [`stackable_certs::ca::Error`] which only implements
+/// [`PartialEq`] for tests.
+#[cfg(test)]
+impl PartialEq for Error {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (
+                Self::BindTcpListener {
+                    source: lhs_source,
+                    socket_addr: lhs_socket_addr,
+                },
+                Self::BindTcpListener {
+                    source: rhs_source,
+                    socket_addr: rhs_socket_addr,
+                },
+            ) => lhs_socket_addr == rhs_socket_addr && lhs_source.kind() == rhs_source.kind(),
+            (lhs, rhs) => lhs == rhs,
+        }
+    }
+}
+
 /// A server which terminates TLS connections and allows clients to commnunicate
 /// via HTTPS with the underlying HTTP router.
 pub struct TlsServer {